Refactor DatePartitionedAvroFileSource to separate out the mechanism of retrieving files that match a watermark from the logic of creating work units.
